fishnet-exporter/requirements.txt
prometheus-client==0.14.1
requests==2.28.1
pyyaml==6.0
schedule==1.1.0
tabulate==0.9.0
colorama==0.4.6
flask==2.2.3
werkzeug==2.2.3
